renaming prd to reg-org and commenting out the auth.attach in loacl.ts
